Basic implementation of InfiniteMPS class (#385)
* updated docstring and fixed some errors in documentation

* adding dmrg

* split up MPS class into BaseMPS and FiniteMPS

* fixed tests to account for non-connected mps

* break edges after applying two-site gate

* removed comments

* added random initialization

* docstrings and comments

* started DMRG implementation

* added mpo.py

* shifting functions araound

* yapf

* docstring

* revert to upstream/master

* moved measurements to BaseMPS

* move __len__ to BaseMPS

* initialize FiniteMPS in canonical form

* global import of FiniteMPS

* fix tests

* add canonicalize flag to FiniteMPS constructor

* addinc docs

* added FiniteMPS.rst

* renaming

* fixing docs

* documentation

* newline

* typing

* working in iMPS

* working on imps

* removed dtype from config.py

* intermediate commit

* removed dtype from shell_backend

* included eigs into backends
currently only numpy backend has it implemented. other backends throw
an error

* added eigs

* added tests for eigs

* linting and typing

* added eigs to tensorflow backend (no implementation)

* docstring

* add dtype argument to all eigs in all backends

* overwrite

* add TMeigs and a bunch of other stuff

* fix tests, add tests

* add dtype argument to eigs

* TMeigs -> transfer_matrix_eigs

tol -> precision

* fixed tests

* pulled in changes from imps branch for backends

* linting

* added eigh test

* added eigh tests

* added eigh test

* added eigh

* added eigh testing

* added eigh

* fix dtype

* added eigh test

* added eigh

* added eigh

* linting

* linting and typing

* added seed to all randn calls

* added tests for index_update

* added index_update

* added test for index_update

* added index_update tests

* added index_update

* return a copy in index_update

* make copy in index_update

* modified tests

* modified tests

* modified tests

* added index_update

* added index_update

* added index_update test

* added index_update

* Revert "added tests for index_update"

This reverts commit 21f7e5aa25acec63c404d90a689adfa01e508705.

* removed index_update test

* added scipy to requirements

* removed unimplemented functions from backends

* raise NotImplementedError with a better message

* added eigs test

* fixed bug in eigs

* fix merge issues

* ncv->num_krylov_vecs

* ncv -> num_krylov_vecs

* ncv -> num_krylov_vecs

* ncv -> num_krylov_vecs

* ncv -> num_krylov_vecs

* ncv -> num_krylov_vecs

* ncv -> num_krylov_vecs

* fix tests

* ncv -> num_krylov_vecs

* num_krylov_vecs -> ncv in scipy.sparse.linalg.eigs

* requirements updated

* scipy1.1

* ncv -> num_krylov_vecs

* revert mps.py and mps_class_test.py to master branch

* removed InfiniteMPS

* add eigh back

* add InfiniteMPS class

* self.backend.name -> self.backend

* mps.py -> infinite_mps.py, finite_mps.py, base_mps.py

* split up tests

* remove mps_class_test.py

* remove imports

* add test

* docstring

* formatting

* remove enable_v2_behaviour() from tests
