+++ tensornetwork/matrixproductstates/__init__.py
@@ -1,1 +1,3 @@
-from tensornetwork.matrixproductstates import mps
+from tensornetwork.matrixproductstates import base_mps
+from tensornetwork.matrixproductstates import finite_mps
+from tensornetwork.matrixproductstates import infinite_mps
